* add specialized streams for 64 bit time stamps   (reduced overhead for time stamps more than 2mins apart) * use 64 bit timestamps throughout the log cache
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/branches/LogCaching@7597 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/LogCache/PackedDWORDInStream.h
@@ -87,7 +87,7 @@ S& operator>> (S& stream, std::vector<V>& data)
 {
 	typedef typename std::vector<V>::const_iterator IT;
 
-	size_t count = stream.GetValue();
+	size_t count = (size_t)stream.GetValue();
 
 	data.clear();
 	data.reserve (count);
